Added ImpHLocalNormal[]
Generalization of the function generating local Hamiltonians for each sector. The new version is more complicated than the previous one: each Hsector is now made by several Hblocks representing the Hubbard interaction in the two orbitals and the interorbital density-density interactions and the energy shift term proportional to the total density.
This resembles the approach used in ImpHblocksNormal[].
For the moment, spin exchange and pair hopping are not included. The idea is to include them in a new function to be committed in the future.