Critical bug fix
The bitwise operations BitOr and BitAnd used to make hopping processes had a critical bug due to a wrong use of BitNot. 
A simpler and more readable implementation has been used to fix the problem, by using BitClear and BitSet at the cost of doubling the evaluation time of hopping functions.