Bug fix: asymmetric Hamiltonian
The local Hamiltonian was giving asymmetric eigenvalues in symmetric sectors with Norb = 2.  This resulted in different Green functions associated to two different orbitals.
For example, with L=3, f=2, Norb=2, EdMode=Normal, the eigenstates of the sectors {2,1,1,1} and {1,1,2,1} were different, even though there is a clear orbital-symmetry between the two sectors.
The reason was a misplaced line of code when initializing the Hblock variable inside Hlocal[]. Easy fix by moving the line to the correct position.