CRITICAL BUGFIX
A critical bug has been corrected.
When the ground state was degenerate, instead of taking the averages of the impurity Green functions computed from each degenerate state, the average of the inverses was taken. This is obviously not the same thing and in fact this was preventing convergence in the presence of degeneracies.