Compatibility fix + bug fix
We no longer use MapApply, which is not compatible with previous versions of Mathematica.
Instead we use Apply[f, args, {1}], which should be equivalent but compatible with previous versions.
Moreover we have made the code compatible with cluster calculation by removing everything that required a front end (just put as a comment so that it can be easily ripristinated). 

Furthermore, there were some bugs related to the new implementation of mixing. We were inverting lists of zeroes for a small bug.
Another bug was inside the observable CdgCdg[], which had the usual sign problem related to swapping of fermionic operators.