Improving pair hopping and Raman mode
The pair hopping is now generalized as follows. The input parameter Jph is no longer a single number, but a Norb x Norb matrix whose diagonal components are intraorbital pair hopping, while off-diagonal terms are interorbital pair hoppings. The matrix is assumed to be symmetric, meaning that the lower triangular part is essentially neglected during the calculation.
A bug on the Hamiltonian construction for the pair hopping part has been fixed. Essentially, a wrong sign factor was implemented.

Added DimSectorList[] for EdMode = "Raman".