Added magnetic field
Warning! This could affect previously working code.
The list InteractionParameters now contains a first item which represents the magnetic field contribution to the Hamiltonian. The term is always present, regardless of EdMode.
The term is a list of Norb x f elements. Each element, labeled by two indexes (orb, sigma), represents the coupling constant in front of a term like n_imp_orb_sigma .
If EdMode == "Raman", then you can provide the Raman hopping matrix M, i.e. a list of Norb matrices with dimension fxf; each of them represents the Raman matrix for a given orbital. In this case the magnetic field is assumed to be the list of eigenvalues of M_orb