Added GetLatticeEnergiesRaman[]
If EdMode = "Raman" the previous paradigm for getting the numeric dispersion relation can't be used, since it implies spin symmetry. In order to avoid a complete change of paradigm, with possible problems on working code, I decided to build a new function that does the same job as GetLatticeEnergies[] but when spin symmetry can't be implied.
The function is GetLatticeEnergiesRaman[] and it assumes that the basis states are the ones that diagonalize the Raman hopping matrix. The function just works for Hypercubic lattice thus far (can't work on Bethe lattice unless gamma=0).