Change paradigm of Raman calculations
The paradigm is changed: instead of using effective flavors that diagonalize the non-interacting Hamiltonian in absence of gauge field, we now work with real flavors. This means that the impurity feels the real Raman hopping scheme instead of an effective magnetic field.
This is apparently working a little bit better than before, and the code appears cleaner in some points.