Bug fixes in spin exchange + implementing Lanczos momenta
Some bugs have been fixed in the Hamiltonian construction for spin exchange and Raman processes. Now the multiorbital atomic limit in presence of a Raman field is consistent with what we expected.
Moreover, the impurity Green function has a tuneable number of "Lanczos momenta": the minimum number of Lanczos iterations for computing Green function is set as an input, and if it's too high, it is automatically reduced to the size of the given Hamiltonian block.