Implementation of EdMode = "Magnetic"
The Raman mode was suitable for off-diagonal spin terms (spin flips), but was not optimized to handle magnetic fields. In the presence of a magnetic field:
- the symmetry operators are all the n_orb_sigma (not only n_orb)
- the Green functions don't have off-diagonal terms, which means that 
   - off diagonal components don't need to be computed
   - inverse functions can be computed faster because it boils down to simple number divisions;

Sector selection.
Providing a specific list of sectors quantum numbers, the user can restrict the diagonalization to those sectors only, with a considerable speedup in situations where the filling factor is under control.

Symbols management
the automatic generation of symbols respecting given symmetries had some bugs, now hopefully fixed.