Improving restricted diagonalization #1
With this update, QnsSectorList and QnsSectorListToDiagonalize are different if the user wants to load the latter from file. This means that E.D. is done only on the sectors specified by QnsSectorListToDiagonalize, while all the sectors are stored.
Next I will fix some bugs (I'm sure I will find some...) and maybe reduce the number of built sectors as well, according to what the user wants.